test: fix testing and tabbing
